src/requirements.txt
APScheduler==3.6.3
paho-mqtt==1.4.0
pymodbus==2.2.0
